clippy: format

+++ src/ui/hud/crosshair.rs
@@ -3,7 +3,7 @@ use bevy::prelude::*;
 
 pub fn setup_crosshair(mut commands: Commands, gui_scale: Res<GUIScale>) {
     let scale: f32 = match *gui_scale {
-        GUIScale::Auto => 2 as f32,
+        GUIScale::Auto => 2_f32,
         GUIScale::Scale(x) => x as f32,
     };
